pass globs as strings when starting the server
Vim uses the shell to execute `system()`, which will expand the globs
for you before executing the command. This means that you will end up
calling `psc-ide-server` passing every purescript file in your source
and bower directories as command line arguments. Aside from the fact
that there is a limit to the argument length for a command, this also
means that if you add one or more files to the project and issue a
`load` command, those source files will not be loaded until the server
is stopped and restarted, meaning you can't jump to definitions.

Fortunately, `psc-ide-server` was built to take source globs, so all we
need to do is quote the globs so that the shell does not expand them. In
addition, the ide server resolves those globs relative to the working
directory, which is specified with `-d`. Since we provide this, we don't
have to concatenate the directory to the globs ourselves.

I did not change the windows command, as I did not have a way of testing
to verify that it works there too.