offer alphanumeric shortcuts for larger options
If you try to use go-to-definition or import for an identifier that
resolves to a few modules, you are offered with a choice where you can
select by the first character of each option. Currently, the index is
prepended to the option so that you can choose more easily, but this
falls short if there are more than 9 options, since the shortcut for
"10" is really "(1)", which conflicts with index 1. This can happen
quite often if you have enough dependencies, especially when modules
re-export a common function.

To get around this, we can start labeling entries with ascii characters
starting with A, B, C... after you reach 9. In a way, it is like base
36, but it would keep going to further ascii characters if there were
enough options. I presume that if that were ever the case, this would be
unusable anyway. This may not be the most glamorous solution, but is
still fairly easy to use.

I noticed that there was an opportunity to refactor to reduce some
duplication between the two code paths, so a new helper function was
introduced that takes a list of dicts plus a key that represents what
the label value is for each item. I also added a space of padding
between the choice and the label because I think it is easier to read.
The return value is wrapped in a dict itself, since it is hard to test
success if we just return either a dict or `v:none`.